Improve responsiveness, modularize home page
