+++ src/App.scss
@@ -13,262 +13,12 @@
 }
 
 .App {
-	font-family: Inter, sans-serif;
-
 	main {
 		.mainContentWrapper {
 			display: flex;
 			flex-direction: column;
-			background-color: $backgroundColor;
-			background-image: url('static/background.png');
-			background-attachment: fixed;
-			background-size: contain;
-
 			width: 100%;
 			height: 100vh;
 		}
-
-		.mainContent {
-			background-color: darken($white-ter, 1%);
-			border-radius: $borderRadius;
-			color: $grey-dark;
-			box-shadow: 4px 4px 4px 2px rgba(0, 0, 0, 0.1);
-			position: relative;
-			margin-top: ($pageMargin / 2);
-			width: calc(100% - calc(var(--page-margin) * 2));
-
-			@media screen and (max-width: $tablet) {
-				margin-top: 0;
-				width: 100%;
-			}
-
-			nav.navbar {
-				border-top-left-radius: $borderRadius;
-				border-top-right-radius: $borderRadius;
-				background-color: $white-ter;
-				padding: 12px;
-
-				.navbar-menu {
-					box-shadow: none !important;
-					background-color: transparent;
-
-					.navbar-item, a.navbar-link {
-						color: $black-bis;
-					}
-				}
-
-				.navbar-start {
-					> .navbar-item {
-						padding: 4px 12px;
-						font-family: 'Noto Sans', sans-serif;
-						font-size: 18px;
-
-						&.has-dropdown {
-							padding-right: 4px;
-						}
-					}
-				}
-
-				.navbar-item:hover, .navbar-item:hover > .navbar-link {
-					background-color: darken($white-ter, 5%);
-					border-radius: $borderRadius;
-				}
-
-				.navbar-dropdown {
-					border-top-width: 3px;
-					border-top-color: $blue;
-					border-bottom-left-radius: $borderRadius;
-					border-bottom-right-radius: $borderRadius;
-					background-color: darken($white-ter, 5%);
-
-					.navbar-item {
-						&:hover {
-							background-color: darken($white-ter, 10%);
-						}
-					}
-				}
-			}
-
-			.heroBoxes {
-				padding-bottom: 0;
-				background-color: darken($grey-darker, 5%);
-				border-top: 6px solid lighten($grey-darker, 10%);
-
-				.heroInfoColumn {
-					position: relative;
-					backdrop-filter: blur(2px);
-					border-radius: $borderRadius;
-					color: $white;
-
-					h2 {
-						color: $white;
-						display: inline-block;
-						margin-top: 0;
-						margin-bottom: 16px;
-					}
-
-					p {
-						margin-left: 16px;
-						font-size: 14px;
-					}
-
-					button {
-						background-color: $black-ter;
-						color: $white-bis;
-						border: 1px solid $blue;
-						border-left: 4px solid $blue;
-						transition: all 200ms ease-in-out;
-						border-radius: $borderRadius;
-
-						&:hover {
-							background-color: darken($black-ter, 3%);
-							color: $white;
-						}
-					}
-
-
-				}
-
-				.heroDelegationInfo {
-					.poolIDRow {
-						margin-left: 16px;
-						display: block;
-						margin-top: 24px;
-						font-size: 14px;
-					}
-
-					.poolIDWrapper {
-						margin-top: 8px;
-					}
-
-					.poolID {
-						font-size: 12px;
-						border: 1px solid #BD7E59;
-						border-left: 4px solid #BD7E59;
-						background-color: $black-ter;
-						color: $white-bis;
-						padding: 6px 6px 6px 12px;
-						border-radius: $borderRadius + 1;
-						transition: all 200ms ease-in-out;
-						max-width: 450px;
-						position: relative;
-						cursor: pointer;
-
-						&:hover {
-							background-color: darken($black-ter, 3%);
-							color: $white;
-						}
-
-						&.copied {
-							border-color: $green;
-						}
-
-						.copyPoolIDBox {
-							display: flex;
-							flex-direction: row;
-							justify-content: center;
-							align-items: center;
-
-							.copyPoolIDText {
-								display: flex;
-								flex-grow: 0;
-								flex-shrink: 1;
-								max-width: 90%;
-								white-space: normal;
-								line-break: anywhere;
-							}
-
-							.copyPoolIDBButton {
-								flex-grow: 0;
-								flex-shrink: 0;
-								display: inline-block;
-								text-align: right;
-								width: 40px;
-								margin-left: 18px;
-								cursor: pointer;
-								user-select: none;
-								-moz-user-select: none;
-								-webkit-user-select: none;
-								transition: all 100ms ease-in-out;
-
-								&:hover {
-									color: $cyan;
-								}
-							}
-
-							.copied {
-								color: $cyan;
-							}
-						}
-					}
-				}
-
-				.payoutInfo {
-					.payoutButton {
-						margin-top: 24px;
-
-						button {
-							outline: none !important;
-							box-shadow: none !important;
-
-							.fa-calendar {
-								margin-right: 8px;
-							}
-
-							span {
-								font-size: 16px;
-								position: relative;
-								top: -1px;
-								font-variant: all-small-caps;
-							}
-						}
-					}
-				}
-			}
-
-			.columnContent {
-				width: 100%;
-				padding: ($pageMargin / 2) $pageMargin;
-				text-align: left;
-
-				@media screen and (max-width: $tablet) {
-					padding: 6px 12px;
-				}
-			}
-
-			.aboutUsHeader {
-				margin-top: 32px;
-			}
-
-			>.columns {
-				margin: 0;
-			}
-		}
-
-		.poolInformation {
-			background-color: #343a40;
-			padding: 24px;
-			margin-top: $pageMargin;
-			border-radius: $borderRadius;
-
-			h3 {
-				color: $white-ter;
-			}
-		}
-	}
-
-	footer {
-		padding: 12px;
-
-		p {
-			font-size: 12px;
-		}
-	}
-
-	.bullet {
-		position: relative;
-		top: 1px;
-		opacity: 0.75;
-		margin-right: 3px;
 	}
 }
