+++ src/components/common/header/Header.module.scss
@@ -9,8 +9,10 @@
 	box-shadow: 0 3px 2px 2px rgba(0, 0, 0, 0.15);
 	z-index: 10;
 	position: relative;
+	display: flex;
 
-	@media screen and (max-width: $tablet) {
+	@media screen and (max-width: 1580px) {
+		display: block;
 		height: auto;
 	}
 
@@ -72,11 +74,6 @@
 		justify-content: center;
 		align-items: center;
 
-		@media screen and (max-width: $tablet) {
-			margin-top: 12px;
-			position: relative;
-		}
-
 		.headerNav {
 			height: 100%;
 
@@ -95,7 +92,7 @@
 					justify-content: center;
 					height: 100%;
 					padding: 0 16px;
-					font-size: 16px;
+					font-size: 1rem;
 					font-family: Lato, sans-serif;
 					font-weight: 700;
 					color: $woodlandYellow;
@@ -110,6 +107,21 @@
 							color: darken($woodlandYellow, 10%);
 						}
 					}
+				}
+			}
+		}
+
+		@media screen and (max-width: 1580px) {
+			margin-top: 12px;
+			position: relative;
+			align-items: flex-start;
+			width: 100%;
+
+			.headerNav {
+				width: 100%;
+
+				ul {
+					justify-content: flex-start;
 				}
 			}
 		}
