+++ src/index.scss
@@ -12,4 +12,6 @@ body {
 	margin: 0;
 	-webkit-font-smoothing: antialiased;
 	-moz-osx-font-smoothing: grayscale;
+	font-size: 16pt;
+	font-family: Inter, sans-serif;
 }
