+++ src/pages/home/units/footer/FooterUnit.js
@@ -1,0 +1,11 @@
+import styles from "./FooterUnit.module.scss";
+
+export function FooterUnit() {
+	return (
+		<div className={styles.footerUnit}>
+			<p>
+				<strong>Disclaimer</strong>: No one associated with Woodland Pools is a financial advisor, broker, or dealer, and none of the information presented should be construed as financial advice. The information presented on our channel, website, and social media is for entertainment and informational purposes only, and all community members should independently verify and research all information when creating their own investment plans.
+			</p>
+		</div>
+	)
+}
