+++ src/pages/tools/Tools.module.scss
@@ -111,7 +111,7 @@ $toolsHeight: calc(100vh - #{$headerHeight});
 
 				h1 {
 					height: 30px;
-					font-size: 24px;
+					font-size: 1.8rem;
 					display: flex;
 					align-items: center;
 				}
