+++ src/style/common.scss
@@ -2,14 +2,14 @@
 
 p {
 	margin: 12px 0;
-	font-size: 16px;
+	font-size: 1rem;
 	font-family: Inter, sans-serif;
 	color: $textColor;
 	font-weight: 300;
 }
 
 h1 {
-	font-size: 32px;
+	font-size: 2rem;
 	font-weight: bold;
 	margin-top: 8px;
 	margin-bottom: 24px;
@@ -20,7 +20,7 @@ h1 {
 }
 
 h2 {
-	font-size: 24px;
+	font-size: 1.5rem;
 	font-weight: bold;
 	margin-top: 8px;
 	margin-bottom: 24px;
@@ -66,7 +66,7 @@ h3 {
 	}
 
 	.material-icons {
-		font-size: 14px;
+		font-size: 0.85rem;
 		position: relative;
 		top: 2px;
 		margin-left: 2px;
@@ -89,7 +89,7 @@ button {
 }
 
 small {
-	font-size: 12px;
+	font-size: 0.9rem;
 }
 
 .adaSign {
