--- src/style/ribbon.scss
@@ -1,41 +1,0 @@
-.corner-ribbon{
-	width: 100px;
-	background: #e43;
-	position: absolute;
-	top: 6px;
-	left: -32px;
-	text-align: center;
-	letter-spacing: 1px;
-	color: #f0f0f0;
-	transform: rotate(-45deg);
-	-webkit-transform: rotate(-45deg);
-	font-family: 'Noto Sans', cursive, sans-serif;
-	font-size: 16px;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
-
-.corner-ribbon.shadow{
-	box-shadow: 0 0 3px rgba(0,0,0,.3);
-}
-
-/* Different positions */
-
-.corner-ribbon.top-left{
-	top: 6px;
-	left: -32px;
-	transform: rotate(-45deg);
-	-webkit-transform: rotate(-45deg);
-}
-
-.corner-ribbon.white{background: #f0f0f0; color: #555;}
-.corner-ribbon.black{background: #333;}
-.corner-ribbon.grey{background: #999;}
-.corner-ribbon.blue{background: #39d;}
-.corner-ribbon.green{background: #2c7;}
-.corner-ribbon.turquoise{background: #1b9;}
-.corner-ribbon.purple{background: #95b;}
-.corner-ribbon.red{background: #e43;}
-.corner-ribbon.orange{background: #e82;}
-.corner-ribbon.yellow{background: #ec0;}
